// apps/web/src/api/generated/schemas/routersPaymentsPaymentRecordCreate.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { RoutersPaymentsPaymentRecordCreateSaleId } from './routersPaymentsPaymentRecordCreateSaleId';
import type { RoutersPaymentsPaymentRecordCreatePromissoryNoteId } from './routersPaymentsPaymentRecordCreatePromissoryNoteId';
import type { RoutersPaymentsPaymentRecordCreatePaymentDate } from './routersPaymentsPaymentRecordCreatePaymentDate';
import type { RoutersPaymentsPaymentRecordCreateDueDate } from './routersPaymentsPaymentRecordCreateDueDate';
import type { RoutersPaymentsPaymentRecordCreateReferenceNumber } from './routersPaymentsPaymentRecordCreateReferenceNumber';
import type { RoutersPaymentsPaymentRecordCreateNotes } from './routersPaymentsPaymentRecordCreateNotes';
import type { RoutersPaymentsPaymentRecordCreateBranchId } from './routersPaymentsPaymentRecordCreateBranchId';
import type { RoutersPaymentsPaymentRecordCreateTenantId } from './routersPaymentsPaymentRecordCreateTenantId';

export interface RoutersPaymentsPaymentRecordCreate {
  partyId: string;
  saleId?: RoutersPaymentsPaymentRecordCreateSaleId;
  promissoryNoteId?: RoutersPaymentsPaymentRecordCreatePromissoryNoteId;
  amount: number;
  paymentMethod: string;
  paymentType?: string;
  paymentDate?: RoutersPaymentsPaymentRecordCreatePaymentDate;
  dueDate?: RoutersPaymentsPaymentRecordCreateDueDate;
  referenceNumber?: RoutersPaymentsPaymentRecordCreateReferenceNumber;
  notes?: RoutersPaymentsPaymentRecordCreateNotes;
  branchId?: RoutersPaymentsPaymentRecordCreateBranchId;
  tenantId?: RoutersPaymentsPaymentRecordCreateTenantId;
}